adrv9009 headless cleanup: isolated clocking
This is part of an attempt to simplify headless.c
