+++ projects/adrv9009/src.mk
@@ -15,6 +15,7 @@
 #PROFILE = tx_bw200_ir245p76_rx_bw200_or245p76_orx_bw200_or245p76_dc245p76
 PROFILE = tx_bw400_ir491p52_rx_bw200_or245p76_orx_bw400_or491p52_dc245p76
 SRCS := $(PROJECT)/src/app/headless.c					\
+	$(PROJECT)/src/app/app_clocking.c					\
 	$(PROJECT)/src/devices/ad9528/ad9528.c				\
 	$(PROJECT)/src/devices/adi_hal/no_os_hal.c			\
 	$(DRIVERS)/rf-transceiver/talise/api/talise_agc.c			\
@@ -44,6 +45,7 @@ SRCS +=	$(PLATFORM_DRIVERS)/axi_io.c					\
 	$(PLATFORM_DRIVERS)/gpio.c					\
 	$(PLATFORM_DRIVERS)/delay.c
 INCS :=	$(PROJECT)/src/app/app_config.h					\
+	$(PROJECT)/src/app/app_clocking.h						\
 	$(PROJECT)/src/devices/ad9528/ad9528.h				\
 	$(PROJECT)/src/devices/ad9528/t_ad9528.h			\
 	$(PROJECT)/src/devices/adi_hal/adi_hal.h			\
