+++ projects/adrv9009/src/app/headless.c
@@ -27,25 +27,24 @@
 #include "delay.h"
 #include "parameters.h"
 #include "util.h"
-#include "ad9528.h"
-#ifdef ALTERA_PLATFORM
-#include "clk_altera_a10_fpll.h"
-#include "altera_adxcvr.h"
-#else
-#include "xil_cache.h"
-#include "clk_axi_clkgen.h"
-#include "axi_adxcvr.h"
-#endif
 #include "axi_jesd204_rx.h"
 #include "axi_jesd204_tx.h"
 #include "axi_dac_core.h"
 #include "axi_adc_core.h"
 #include "axi_dmac.h"
-
+#ifdef ALTERA_PLATFORM
+#include "altera_adxcvr.h"
+#else
+#include "axi_adxcvr.h"
+#endif
+#ifndef ALTERA_PLATFORM
+#include "xil_cache.h"
+#endif
 #include "talise_config_ad9528.h"
 #include "talise_arm_binary.h"
 #include "talise_stream_binary.h"
 #include "app_config.h"
+#include "app_clocking.h"
 
 /**********************************************************/
 /**********************************************************/
@@ -55,89 +54,6 @@
 
 int main(void)
 {
-	adiHalErr_t ad9528Error;
-
-	ad9528Device_t clockAD9528_ = {
-		NULL,
-		NULL,
-		&clockSpiSettings,
-		&clockPll1Settings,
-		&clockPll2Settings,
-		&clockOutputSettings,
-		&clockSysrefSettings,
-		NULL,
-		NULL
-	};
-	ad9528Device_t *clockAD9528_device = &clockAD9528_;
-#ifdef ALTERA_PLATFORM
-	struct altera_spi_init_param ad9528_spi_param = {
-		.type = NIOS_II_SPI,
-		.device_id = 0,
-		.base_address = SPI_BASEADDR
-	};
-	clockAD9528_.extra_spi = &ad9528_spi_param;
-	struct altera_gpio_init_param ad9528_gpio_param = {
-		.type = NIOS_II_GPIO,
-		.device_id = 0,
-		.base_address = GPIO_BASEADDR
-	};
-	clockAD9528_.extra_gpio = &ad9528_gpio_param;
-	struct altera_a10_fpll_init rx_device_clk_pll_init = {
-		"rx_device_clk_pll",
-		RX_A10_FPLL_BASEADDR,
-		clockAD9528_device->outputSettings->outFrequency_Hz[1]
-	};
-	struct altera_a10_fpll_init tx_device_clk_pll_init = {
-		"tx_device_clk_pll",
-		TX_A10_FPLL_BASEADDR,
-		clockAD9528_device->outputSettings->outFrequency_Hz[1]
-	};
-	struct altera_a10_fpll_init rx_os_device_clk_pll_init = {
-		"rx_os_device_clk_pll",
-		RX_OS_A10_FPLL_BASEADDR,
-		clockAD9528_device->outputSettings->outFrequency_Hz[1]
-	};
-	struct altera_a10_fpll *rx_device_clk_pll;
-	struct altera_a10_fpll *tx_device_clk_pll;
-	struct altera_a10_fpll *rx_os_device_clk_pll;
-#else
-	struct xil_spi_init_param ad9528_spi_param = {
-#ifdef PLATFORM_MB
-		.type = SPI_PL,
-#else
-		.type = SPI_PS,
-#endif
-		.device_id = 0
-	};
-	clockAD9528_.extra_spi = &ad9528_spi_param;
-	struct xil_gpio_init_param ad9528_gpio_param = {
-#ifdef PLATFORM_MB
-		.type = GPIO_PL,
-#else
-		.type = GPIO_PS,
-#endif
-		.device_id = GPIO_DEVICE_ID,
-	};
-	clockAD9528_.extra_gpio = &ad9528_gpio_param;
-	struct axi_clkgen_init rx_clkgen_init = {
-		"rx_clkgen",
-		RX_CLKGEN_BASEADDR,
-		clockAD9528_device->outputSettings->outFrequency_Hz[1]
-	};
-	struct axi_clkgen_init tx_clkgen_init = {
-		"tx_clkgen",
-		TX_CLKGEN_BASEADDR,
-		clockAD9528_device->outputSettings->outFrequency_Hz[1]
-	};
-	struct axi_clkgen_init rx_os_clkgen_init = {
-		"rx_os_clkgen",
-		RX_OS_CLKGEN_BASEADDR,
-		clockAD9528_device->outputSettings->outFrequency_Hz[1]
-	};
-	struct axi_clkgen *rx_clkgen;
-	struct axi_clkgen *tx_clkgen;
-	struct axi_clkgen *rx_os_clkgen;
-#endif
 	uint32_t rx_lane_rate_khz = talInit.rx.rxProfile.rxOutputRate_kHz *
 				    talInit.jesd204Settings.framerA.M * (20 /
 						    hweight8(talInit.jesd204Settings.framerA.serializerLanesEnabled));
@@ -150,6 +66,7 @@ int main(void)
 				       talInit.jesd204Settings.framerB.M * (20 /
 						       hweight8(talInit.jesd204Settings.framerB.serializerLanesEnabled));
 	uint32_t rx_os_div40_rate_hz = rx_os_lane_rate_khz * (1000 / 40);
+
 	struct jesd204_rx_init rx_jesd_init = {
 		"rx_jesd",
 		RX_JESD_BASEADDR,
@@ -346,123 +263,7 @@ int main(void)
 	/**********************************************************/
 	/**********************************************************/
 
-	/** < Insert User System Clock(s) Initialization Code Here >
-	 * System Clock should provide a device clock and SYSREF signal
-	 * to the Talise device.
-	 **/
-	/* Init the AD9528 data structure */
-	ad9528Error = AD9528_initDeviceDataStruct(clockAD9528_device,
-			clockAD9528_device->pll1Settings->vcxo_Frequency_Hz,
-			clockAD9528_device->pll1Settings->refA_Frequency_Hz,
-			clockAD9528_device->outputSettings->outFrequency_Hz[1]);
-	if (ad9528Error != ADIHAL_OK) {
-		printf("error: AD9528_initDeviceDataStruct() failed\n");
-		goto error_0;
-	}
-
-	/* Perform a hard reset on the AD9528 DUT */
-	ad9528Error = AD9528_resetDevice(clockAD9528_device);
-	if (ad9528Error != ADIHAL_OK) {
-		printf("error: AD9528_resetDevice() failed\n");
-		goto error_1;
-	}
-
-	/* Initialize the AD9528 by writing all SPI registers */
-	ad9528Error = AD9528_initialize(clockAD9528_device);
-	if (ad9528Error == ADIHAL_ERR) {
-		printf("error: AD9528_initialize() failed\n");
-		goto error_1;
-	}
-
-	if (ad9528Error == ADIHAL_WARNING)
-		printf("warning: AD9528_initialize() issues. "
-		       "Possible cause: REF_CLK not connected\n");
-
-#ifdef ALTERA_PLATFORM
-	/* Initialize A10 FPLLs */
-	status = altera_a10_fpll_init(&rx_device_clk_pll,
-				      &rx_device_clk_pll_init);
-	if (status != SUCCESS) {
-		printf("error: %s: altera_a10_fpll_init() failed\n",
-		       rx_os_device_clk_pll_init.name);
-		goto error_1;
-	}
-	status = altera_a10_fpll_init(&tx_device_clk_pll,
-				      &tx_device_clk_pll_init);
-	if (status != SUCCESS) {
-		printf("error: %s: altera_a10_fpll_init() failed\n",
-		       rx_os_device_clk_pll_init.name);
-		goto error_2;
-	}
-	status = altera_a10_fpll_init(&rx_os_device_clk_pll,
-				      &rx_os_device_clk_pll_init);
-	if (status != SUCCESS) {
-		printf("error: %s: altera_a10_fpll_init() failed\n",
-		       rx_os_device_clk_pll_init.name);
-		goto error_3;
-	}
-
-	altera_a10_fpll_disable(rx_device_clk_pll);
-	status = altera_a10_fpll_set_rate(rx_device_clk_pll,
-					  rx_div40_rate_hz);
-	if (status != SUCCESS) {
-		printf("error: %s: altera_a10_fpll_set_rate() failed\n",
-		       rx_device_clk_pll->name);
-		goto error_4;
-	}
-	altera_a10_fpll_enable(rx_device_clk_pll);
-	altera_a10_fpll_disable(tx_device_clk_pll);
-	status = altera_a10_fpll_set_rate(tx_device_clk_pll,
-					  tx_div40_rate_hz);
-	if (status != SUCCESS) {
-		printf("error: %s: altera_a10_fpll_set_rate() failed\n",
-		       tx_device_clk_pll->name);
-		goto error_4;
-	}
-	altera_a10_fpll_enable(tx_device_clk_pll);
-	altera_a10_fpll_disable(rx_os_device_clk_pll);
-	status = altera_a10_fpll_set_rate(rx_os_device_clk_pll,
-					  rx_os_div40_rate_hz);
-	if (status != SUCCESS) {
-		printf("error: %s: altera_a10_fpll_set_rate() failed\n",
-		       rx_os_device_clk_pll->name);
-		goto error_4;
-	}
-	altera_a10_fpll_enable(rx_os_device_clk_pll);
-#else
-	/* Initialize CLKGEN */
-	status = axi_clkgen_init(&rx_clkgen, &rx_clkgen_init);
-	if (status != SUCCESS) {
-		printf("error: %s: axi_clkgen_init() failed\n", rx_clkgen_init.name);
-		goto error_1;
-	}
-	status = axi_clkgen_init(&tx_clkgen, &tx_clkgen_init);
-	if (status != SUCCESS) {
-		printf("error: %s: axi_clkgen_init() failed\n", tx_clkgen_init.name);
-		goto error_2;
-	}
-	status = axi_clkgen_init(&rx_os_clkgen, &rx_os_clkgen_init);
-	if (status != SUCCESS) {
-		printf("error: %s: axi_clkgen_set_rate() failed\n", rx_os_clkgen_init.name);
-		goto error_3;
-	}
-
-	status = axi_clkgen_set_rate(rx_clkgen, rx_div40_rate_hz);
-	if (status != SUCCESS) {
-		printf("error: %s: axi_clkgen_set_rate() failed\n", rx_clkgen->name);
-		goto error_4;
-	}
-	status = axi_clkgen_set_rate(tx_clkgen, tx_div40_rate_hz);
-	if (status != SUCCESS) {
-		printf("error: %s: axi_clkgen_set_rate() failed\n", tx_clkgen->name);
-		goto error_4;
-	}
-	status = axi_clkgen_set_rate(rx_os_clkgen, rx_os_div40_rate_hz);
-	if (status != SUCCESS) {
-		printf("error: %s: axi_clkgen_set_rate() failed\n", rx_os_clkgen->name);
-		goto error_4;
-	}
-#endif
+	clocking_init(rx_div40_rate_hz, tx_div40_rate_hz, rx_os_div40_rate_hz);
 
 	/*** < Insert User BBIC JESD204B Initialization Code Here > ***/
 
@@ -470,17 +271,17 @@ int main(void)
 	status = axi_jesd204_rx_init(&rx_jesd, &rx_jesd_init);
 	if (status != SUCCESS) {
 		printf("error: %s: axi_jesd204_rx_init() failed\n", rx_jesd_init.name);
-		goto error_4;
+		goto error_5;
 	}
 	status = axi_jesd204_tx_init(&tx_jesd, &tx_jesd_init);
 	if (status != SUCCESS) {
 		printf("error: %s: axi_jesd204_tx_init() failed\n", tx_jesd_init.name);
-		goto error_5;
+		goto error_6;
 	}
 	status = axi_jesd204_rx_init(&rx_os_jesd, &rx_os_jesd_init);
 	if (status != SUCCESS) {
 		printf("error: %s: axi_jesd204_rx_init() failed\n", rx_os_jesd_init.name);
-		goto error_6;
+		goto error_7;
 	}
 
 	/* Initialize ADXCR */
@@ -896,16 +697,7 @@ int main(void)
 	axi_jesd204_rx_remove(rx_os_jesd);
 	axi_jesd204_tx_remove(tx_jesd);
 	axi_jesd204_rx_remove(rx_jesd);
-#ifdef ALTERA_PLATFORM
-	altera_a10_fpll_remove(rx_device_clk_pll);
-	altera_a10_fpll_remove(tx_device_clk_pll);
-	altera_a10_fpll_remove(rx_os_device_clk_pll);
-#else
-	axi_clkgen_remove(rx_os_clkgen);
-	axi_clkgen_remove(tx_clkgen);
-	axi_clkgen_remove(rx_clkgen);
-#endif
-	AD9528_remove(clockAD9528_device);
+	clocking_deinit();
 	printf("Bye\n");
 
 	return SUCCESS;
@@ -924,26 +716,6 @@ error_6:
 	axi_jesd204_tx_remove(tx_jesd);
 error_5:
 	axi_jesd204_rx_remove(rx_jesd);
-error_4:
-#ifdef ALTERA_PLATFORM
-	altera_a10_fpll_remove(rx_os_device_clk_pll);
-#else
-	axi_clkgen_remove(rx_os_clkgen);
-#endif
-error_3:
-#ifdef ALTERA_PLATFORM
-	altera_a10_fpll_remove(tx_device_clk_pll);
-#else
-	axi_clkgen_remove(tx_clkgen);
-#endif
-error_2:
-#ifdef ALTERA_PLATFORM
-	altera_a10_fpll_remove(rx_device_clk_pll);
-#else
-	axi_clkgen_remove(rx_clkgen);
-#endif
-error_1:
-	AD9528_remove(clockAD9528_device);
 error_0:
 	return FAILURE;
 }
